requirements.txt
aiogram_dialog
aiogram
python-decouple
sqlalchemy
asyncpg
llamaapi
openai